Prefix google extractor name with "edfi_"
