+++ src/google-classroom-extractor/.mypy.ini
@@ -1,5 +1,5 @@
 [mypy]
-files=google_classroom_extractor
+files=edfi_google_classroom_extractor
 
 [mypy-pandas]
 ignore_missing_imports = True
