+++ src/google-classroom-extractor/tests/api/courses/test_courses_request.py
@@ -6,7 +6,7 @@
 import pytest
 import pook
 from pandas.core.frame import DataFrame
-from google_classroom_extractor.api.courses import request_latest_courses_as_df
+from edfi_google_classroom_extractor.api.courses import request_latest_courses_as_df
 from tests.api.api_helper import setup_fake_classroom_api
 
 # unique value for each column in fixture
