+++ src/google-classroom-extractor/tests/mapping/sections/test_sections.py
@@ -4,8 +4,8 @@
 # See the LICENSE and NOTICES files in the project root for more information.
 import pytest
 from pandas import DataFrame
-from google_classroom_extractor.mapping.sections import courses_to_sections_df
-from google_classroom_extractor.mapping.constants import SOURCE_SYSTEM
+from edfi_google_classroom_extractor.mapping.sections import courses_to_sections_df
+from edfi_google_classroom_extractor.mapping.constants import SOURCE_SYSTEM
 
 # unique value for each column in fixture
 COURSE_ID = "1"
